fix(Event): prefix GameObject component to prevent overriding
There was an issue where if the GameObject component name was the
same as an actual component on the GameObject then the GameObject
would override the component object and only the GameObject would
show up in the list.

This has been fixed by simply prefixing the GameObject element
in the list with the word "GameObect (<actual name>)" to ensure
uniqueness.